refactor: list menu items imports & exports
