+++ src/components/UI/molecules/ListMenuItems/index.ts
@@ -1,1 +1,2 @@
+export * from './ListMenuItems.interface'
 export { ListMenuItems } from './ListMenuItems.component'
